+++ app/src/main/java/com/hongyuan/fitness/ui/person/person_message/PersonMessageBeans.java
@@ -42,7 +42,7 @@ public class PersonMessageBeans extends BaseBean {
         private int pid;
         private int cid;
         private String mi_sign;
-        private int mi_birth;
+        private long mi_birth;
         private int gz_num;
         private int fs_num;
         private String birth;
@@ -122,11 +122,11 @@ public class PersonMessageBeans extends BaseBean {
             this.mi_sign = mi_sign;
         }
 
-        public int getMi_birth() {
+        public long getMi_birth() {
             return mi_birth;
         }
 
-        public void setMi_birth(int mi_birth) {
+        public void setMi_birth(long mi_birth) {
             this.mi_birth = mi_birth;
         }
